Restart workspace library scanner when triggered
Until now, the workspace library scanner was not restarted on every
trigger, thus triggers while the scan was running were lost, which lead
to outdated information in the database. Now the scan is immediately
restarted if it is triggered while already running. This avoids outdated
information, and is important to keep the list of libraries (especially
in the library manager) always up to date when libraries are added or
removed.