Refactor and improve scanning of workspace libraries
WorkspaceLibraryScanner now also scans which libraries are installed
instead of using the libraries loaded by the Workspace object. This
allows to add/remove/update workspace libraries without notifying the
Workspace object about the exact changes. To get faster feedback of the
changes (e.g. to update the widgets of the library manager), the scanner
now emits a signal after the library list is updated.

In addition, the scanner now emits a new signal scanFinished() which is
used in many classes to get notified even if the scan fails or was
aborted.

Also the debug output is now a bit more verbose to see more information
about the library scan.